requirements.txt
appdirs~=1.4.0
click~=6.7
Flask~=0.12
geopandas~=0.2.1
gunicorn~=19.6.0
itsdangerous~=0.24
Jinja2~=2.9.5
MarkupSafe~=0.23
packaging~=16.8
pyparsing~=2.1.10
pytest~=3.0.6
rtree~=0.8.3
six~=1.10.0
Werkzeug~=0.11.15
